Restructuring the plugins, part IV: - New plugin set linuxdevices/nieseries - Fixed VPATH build problem with DX_CONFIG (removed $(srdir))
git-svn-id: https://svn.code.sf.net/p/relacs/code/relacs/trunk@209 1550813e-0d35-0410-aec7-fc86b090456a

+++ release.sh
@@ -3,20 +3,33 @@
 if test "x$1" = "x--help"; then
   echo "Usage:"
   echo ""
-  echo "release.sh <install-path> <configure-options>"
+  echo "release.sh [-l] [-d] <install-path> <configure-options>"
   echo ""
   echo "Calls ./bootstrap.sh && ./configure && make && make check &&"
   echo "make install && make distcheck for the whole relacs project"
-  echo "and the individual relacs libraries."
+  echo "and the individual relacs libraries (if the -l option is supplied)."
   echo "You must supply an absolute <install-path> as the first argument."
   echo "All following arguments are supplied as further options to configure."
   echo ""
+  echo "-d: make distcheck only"
   exit 0
+fi
+
+CHECKLIBS=no
+if test "x$1" = "x-l"; then
+  CHECKLIBS=yes
+  shift
+fi
+
+DISTCHECKONLY=no
+if test "x$1" = "x-d"; then
+  DISTCHECKONLY=yes
+  shift
 fi
 
 INSTALLPATH="$1"
 shift
-if [ x${INSTALLPATH} = x ]; then
+if test "x${INSTALLPATH}" = x; then
   echo "You need to specify an absolute installation path."
   exit 1
 fi
@@ -24,8 +37,9 @@ CONFIGUREPREFIX="--prefix=$INSTALLPATH"
 
 # further options for configure (without prefix):
 CONFIGUREFLAGS="$@"
+export DISTCHECK_CONFIGURE_FLAGS="$CONFIGUREFLAGS"
 
-if [ x${MAKE} = x ]; then
+if test "x${MAKE}" = x; then
     MAKE=make
 fi
 
@@ -38,37 +52,44 @@ echo ""
 echo "Check relacs:"
 ./bootstrap.sh || exit 1
 ./configure $CONFIGUREPREFIX $CONFIGUREFLAGS # Also creates configure.ac files of libraries
-${MAKE} || exit 1
-${MAKE} check || exit 1
-${MAKE} install || exit 1
+if test $DISTCHECKONLY = no; then
+  ${MAKE} || exit 1
+  ${MAKE} check || exit 1
+  ${MAKE} install || exit 1
+fi
 ${MAKE} distcheck || exit 1
 echo ""
 echo "Successfully checked relacs"
 echo ""
 
 
-# configure and check everything:
-export CPPFLAGS="-I${INSTALLPATH}/include"
-export LDFLAGS="-L${INSTALLPATH}/lib"
-export DISTCHECK_CONFIGURE_FLAGS="$CONFIGUREFLAGS"
-for i in numerics daq options datafile widgets config plot; do
-    echo "Check $i library:"
-    cd $i || exit 1
-    find . -name '*.la' -print -delete
-    ./bootstrap.sh || exit 1
-    ./configure $CONFIGUREPREFIX $CONFIGUREFLAGS || exit 1
-    ${MAKE} || exit 1
-    ${MAKE} check || exit 1
-    ${MAKE} install || exit 1
-    ${MAKE} distcheck || exit 1
-    cd - || exit 1
-    echo ""
-    echo "Successfully checked $i library"
-    echo ""
-done
+# configure and check everything for the libraries:
+if test $CHECKLIBS = yes; then
+  export CPPFLAGS="-I${INSTALLPATH}/include"
+  export LDFLAGS="-L${INSTALLPATH}/lib"
+  for i in numerics daq options datafile widgets config plot; do
+      echo "Check $i library:"
+      cd $i || exit 1
+      find . -name '*.la' -print -delete
+      ./bootstrap.sh || exit 1
+      ./configure $CONFIGUREPREFIX $CONFIGUREFLAGS || exit 1
+      if test $DISTCHECKONLY = no; then
+        ${MAKE} || exit 1
+        ${MAKE} check || exit 1
+        ${MAKE} install || exit 1
+        ${MAKE} distcheck || exit 1
+      fi
+      cd - || exit 1
+      echo ""
+      echo "Successfully checked $i library"
+      echo ""
+  done
 
-echo "Successfully checked relacs and all its libraries!"
-echo "You can go ahead and make a release (using releaseforge) :-)"
-echo ""
+  echo "Successfully checked relacs and all its libraries!"
+  echo "You can go ahead and make a release (using releaseforge) :-)"
+  echo ""
+
+fi
+
 exit 0
